requirements.txt
pymongo
python-dotenv
google-genai
streamlit
numpy
scikit-learn
umap-learn